Subzero: Remove the need for a separate NaCl SDK installation.
Now it assumes tests are being run from within the Subzero portion of the native_client tree, and sets up PATH to find PNaCl tools relative to there.

It may be necessary to do a one-time setup to be able to build pexes:

    pnacl/build.sh sdk newlib

or the equivalent scons commands.

If the tool chain is updated, propagate the changes via:

    toolchain_build/toolchain_build_pnacl.py llvm_i686_linux --install=toolchain/linux_x86/pnacl_newlib

BUG= none
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/525603002