Subzero: Always be sure x87 FP stack is emptied after a call.
In x86-32, floating point values are returned to the caller on the top of the x87 floating point stack.  The caller is required to remove it from the x87 FP stack, e.g. via the fstp instruction.

This must be done even when the return value is not actually used anywhere else in the function, in which case O2 is likely to want to dead-code eliminate the fstp instruction.

We enforce this by adding a fake use of the fstp destination.

BUG= none
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/563303003