Subzero: Make the szbuild.py script more automatic.
One wants to omit the --init option to make things go faster, but then there's the risk that the pexe or llvm2ice or llc has changed and --init is actually necessary.

This change automatically compares the modification timestamps of the pexe and the llvm2ice and llc binaries against the modification timestamps of the object files to determine whether re-translation is needed.

The --init option forces re-translation regardless of timestamps.

BUG= none
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/551373008