Handle "Mov" which is mov, movss, movsd, and used for nacl.read.tp.
Currently, this only checks and emits the segment
override only for GPR instructions, assuming it's mostly
only used for nacl.read.tp. The code will assert when used
in other situations.

The lea hack is still tested in some files, but it's not
emitted with emitIAS, and instead the "immediate" operand
now has a fixup.

There is a more compact encoding for "mov eax, moffs32",
etc., but that isn't used right now.

BUG=none
R=stichnot@chromium.org

Review URL: https://codereview.chromium.org/649463002