Fix handling of relocation names, so that prefix mangling works.
Fixes bug in the representation of relocation names (in either global
initializers or as constant expressions in code) so that they
understand when the name is externally defined. This allows us to test
this property using command line arguments, and fixes relocation tests
in cross compilations (where externnally referenced names shouldn't be
name mangled).

BUG=
R=jvoung@chromium.org, stichnot@chromium.org

Review URL: https://codereview.chromium.org/667763002